refactor the config setup
